refactor: organizing files, functions and folders
