+++ tests/test_e2e.py
@@ -5,36 +5,6 @@ from dotenv import load_dotenv
 from .conftest import APP_URL, TOKEN
 
 load_dotenv()
-
-
-def test_login_com_token_incorreto(page: Page):
-    fake_token = secrets.token_hex(16)
-    print(f"[DEBUG] Token gerado para teste: {fake_token}")
-
-    print(f"[DEBUG] Acessando {APP_URL}...")
-    page.goto(APP_URL)
-
-    print("[DEBUG] Aguardando input de token aparecer...")
-    page.wait_for_selector('input[name="token"]', timeout=5000)
-
-    print("[DEBUG] Preenchendo o token falso...")
-    page.fill('input[name="token"]', fake_token)
-
-    print("[DEBUG] Clicando no botão Entrar...")
-    page.click('button[type="submit"]')
-
-    print("[DEBUG] Aguardando mensagem de erro aparecer...")
-    page.wait_for_selector("p.text-red-500", timeout=5000)
-
-    print("[DEBUG] Capturando a mensagem de erro...")
-    error_message = page.text_content("p.text-red-500")
-    print(f"[DEBUG] Mensagem encontrada: {error_message}")
-
-    assert "Token incorreto" in error_message, (
-        f"Mensagem de erro inesperada: {error_message}"
-    )
-
-
 def test_login_com_token_correto(page: Page):
     print(f"[DEBUG] Acessando {APP_URL}...")
     page.goto(APP_URL)
